+++ java-src/keypin/PropertyFile.java
@@ -15,6 +15,7 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.regex.Pattern;
@@ -203,14 +204,22 @@ public class PropertyFile {
         }
     }
 
-    public static Properties resolveConfig(final String configFile, final String parentKey, final Logger logger)
+    public static Properties resolveConfig(final List<String> configFiles, final String parentKey, final Logger logger)
             throws IOException {
-        return realize(readCascadingConfig(configFile, parentKey, logger), logger);
+        final Properties result = new Properties();
+        for (String eachFile: configFiles) {
+            result.putAll(readCascadingConfig(eachFile, parentKey, logger));
+        }
+        return realize(result, logger);
     }
 
-    public static Properties resolveConfig(final String configFile, final Logger logger)
+    public static Properties resolveConfig(final List<String> configFiles, final Logger logger)
             throws IOException {
-        return realize(readConfig(configFile, logger), logger);
+        final Properties result = new Properties();
+        for (String eachFile: configFiles) {
+            result.putAll(readConfig(eachFile, logger));
+        }
+        return realize(result, logger);
     }
 
 }